Polish code for #8879

+++ catalogs/catalog-generic-lakehouse/src/main/java/org/apache/gravitino/catalog/lakehouse/lance/LanceCatalogOperations.java
@@ -19,6 +19,8 @@
 
 package org.apache.gravitino.catalog.lakehouse.lance;
 
+import static org.apache.gravitino.Entity.EntityType.TABLE;
+
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import com.lancedb.lance.Dataset;
@@ -44,7 +46,9 @@ import org.apache.gravitino.EntityStore;
 import org.apache.gravitino.GravitinoEnv;
 import org.apache.gravitino.NameIdentifier;
 import org.apache.gravitino.Namespace;
+import org.apache.gravitino.catalog.lakehouse.GenericLakehouseTablePropertiesMetadata;
 import org.apache.gravitino.catalog.lakehouse.LakehouseCatalogOperations;
+import org.apache.gravitino.catalog.lakehouse.utils.EntityConverter;
 import org.apache.gravitino.connector.CatalogInfo;
 import org.apache.gravitino.connector.GenericLakehouseTable;
 import org.apache.gravitino.connector.HasPropertyMetadata;
@@ -53,9 +57,8 @@ import org.apache.gravitino.exceptions.NoSuchSchemaException;
 import org.apache.gravitino.exceptions.NoSuchTableException;
 import org.apache.gravitino.exceptions.TableAlreadyExistsException;
 import org.apache.gravitino.meta.AuditInfo;
-import org.apache.gravitino.meta.GenericTableEntity;
+import org.apache.gravitino.meta.TableEntity;
 import org.apache.gravitino.rel.Column;
-import org.apache.gravitino.rel.GenericTable;
 import org.apache.gravitino.rel.Table;
 import org.apache.gravitino.rel.TableChange;
 import org.apache.gravitino.rel.expressions.distributions.Distribution;
@@ -70,13 +73,17 @@ import org.apache.hadoop.fs.Path;
 
 public class LanceCatalogOperations implements LakehouseCatalogOperations {
 
+  @SuppressWarnings("UnusedVariable")
   private Map<String, String> lancePropertiesMap;
+
+  private EntityStore store;
 
   @Override
   public void initialize(
       Map<String, String> config, CatalogInfo info, HasPropertyMetadata propertiesMetadata)
       throws RuntimeException {
     lancePropertiesMap = ImmutableMap.copyOf(config);
+    store = GravitinoEnv.getInstance().entityStore();
   }
 
   @Override
@@ -93,13 +100,31 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
 
   @Override
   public NameIdentifier[] listTables(Namespace namespace) throws NoSuchSchemaException {
+    // No need to do anything.
     return new NameIdentifier[0];
   }
 
   @Override
   public Table loadTable(NameIdentifier ident) throws NoSuchTableException {
-    // Should not come here.
-    return null;
+    try {
+      TableEntity tableEntity = store.get(ident, Entity.EntityType.TABLE, TableEntity.class);
+      return GenericLakehouseTable.builder()
+          .withFormat(tableEntity.getFormat())
+          .withProperties(tableEntity.getProperties())
+          .withAuditInfo(tableEntity.auditInfo())
+          .withSortOrders(tableEntity.getSortOrder())
+          .withPartitioning(tableEntity.getPartitions())
+          .withDistribution(tableEntity.getDistribution())
+          .withColumns(EntityConverter.toColumns(tableEntity.columns()))
+          .withIndexes(tableEntity.getIndexes())
+          .withName(tableEntity.name())
+          .withComment(tableEntity.getComment())
+          .build();
+    } catch (NoSuchEntityException e) {
+      throw new NoSuchTableException("No such table: %s", ident);
+    } catch (IOException ioe) {
+      throw new RuntimeException("Failed to load table entity for: " + ident, ioe);
+    }
   }
 
   @Override
@@ -114,7 +139,7 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
       Index[] indexes)
       throws NoSuchSchemaException, TableAlreadyExistsException {
     // Ignore partitions, distributions, sortOrders, and indexes for Lance tables;
-    String location = properties.get("location");
+    String location = properties.get(GenericLakehouseTablePropertiesMetadata.LAKEHOUSE_LOCATION);
     try (Dataset dataset =
         Dataset.create(
             new RootAllocator(),
@@ -158,7 +183,7 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
                                       "",
                                       org.apache.arrow.vector.types.pojo.FieldType.nullable(
                                           childType),
-                                      null))
+                                      null /* child type is null */))
                           .collect(Collectors.toList());
 
                   if (nullable) {
@@ -184,6 +209,7 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
     // Lance only supports adding indexes for now.
     List<Index> addedIndexes = Lists.newArrayList();
 
+    // Only support for adding index for now.
     for (TableChange change : changes) {
       if (change instanceof TableChange.AddIndex addIndexChange) {
         Index index =
@@ -196,17 +222,66 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
       }
     }
 
-    EntityStore entityStore = GravitinoEnv.getInstance().entityStore();
-    GenericTableEntity entity;
+    TableEntity updatedEntity;
     try {
-      entity = entityStore.get(ident, Entity.EntityType.TABLE, GenericTableEntity.class);
+      TableEntity entity = store.get(ident, Entity.EntityType.TABLE, TableEntity.class);
+      updatedEntity =
+          store.update(
+              ident,
+              TableEntity.class,
+              TABLE,
+              tableEntity ->
+                  TableEntity.builder()
+                      .withId(tableEntity.id())
+                      .withName(tableEntity.name())
+                      .withNamespace(tableEntity.namespace())
+                      .withFormat(entity.getFormat())
+                      .withAuditInfo(
+                          AuditInfo.builder()
+                              .withCreator(tableEntity.auditInfo().creator())
+                              .withCreateTime(tableEntity.auditInfo().createTime())
+                              .withLastModifier(PrincipalUtils.getCurrentPrincipal().getName())
+                              .withLastModifiedTime(Instant.now())
+                              .build())
+                      .withColumns(tableEntity.columns())
+                      .withIndexes(
+                          ArrayUtils.addAll(
+                              entity.getIndexes(), addedIndexes.toArray(new Index[0])))
+                      .withDistribution(tableEntity.getDistribution())
+                      .withPartitions(tableEntity.getPartitions())
+                      .withSortOrder(tableEntity.getSortOrder())
+                      .withProperties(tableEntity.getProperties())
+                      .withComment(tableEntity.getComment())
+                      .build());
+
+      // Add indexes to Lance dataset
+      addLanceIndex(updatedEntity, addedIndexes);
+
+      // return the updated table
+      return GenericLakehouseTable.builder()
+          .withFormat(updatedEntity.getFormat())
+          .withProperties(updatedEntity.getProperties())
+          .withAuditInfo(updatedEntity.auditInfo())
+          .withSortOrders(updatedEntity.getSortOrder())
+          .withPartitioning(updatedEntity.getPartitions())
+          .withDistribution(updatedEntity.getDistribution())
+          .withColumns(EntityConverter.toColumns(updatedEntity.columns()))
+          .withIndexes(updatedEntity.getIndexes())
+          .withName(updatedEntity.name())
+          .withComment(updatedEntity.getComment())
+          .build();
     } catch (NoSuchEntityException e) {
       throw new NoSuchTableException("No such table: %s", ident);
     } catch (IOException ioe) {
       throw new RuntimeException("Failed to load table entity for: " + ident, ioe);
     }
+  }
 
-    String location = entity.getProperties().get("location");
+  private void addLanceIndex(TableEntity updatedEntity, List<Index> addedIndexes) {
+    String location =
+        updatedEntity
+            .getProperties()
+            .get(GenericLakehouseTablePropertiesMetadata.LAKEHOUSE_LOCATION);
     try (Dataset dataset = Dataset.open(location, new RootAllocator())) {
       // For Lance, we only support adding indexes, so in fact, we can't handle drop index here.
       for (Index index : addedIndexes) {
@@ -222,28 +297,7 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
             indexParams,
             true);
       }
-    } catch (Exception e) {
-      throw new RuntimeException("Failed to alter Lance table: " + ident, e);
     }
-
-    GenericTable oldTable = entity.toGenericTable();
-    Index[] newIndexes = oldTable.index();
-    for (Index index : addedIndexes) {
-      newIndexes = ArrayUtils.add(newIndexes, index);
-    }
-
-    return GenericLakehouseTable.builder()
-        .withFormat(oldTable.format())
-        .withProperties(oldTable.properties())
-        .withAuditInfo((AuditInfo) oldTable.auditInfo())
-        .withSortOrders(oldTable.sortOrder())
-        .withPartitioning(oldTable.partitioning())
-        .withDistribution(oldTable.distribution())
-        .withColumns(oldTable.columns())
-        .withIndexes(newIndexes)
-        .withName(oldTable.name())
-        .withComment(oldTable.comment())
-        .build();
   }
 
   private IndexParams getIndexParamsByIndexType(IndexType indexType) {
@@ -266,8 +320,17 @@ public class LanceCatalogOperations implements LakehouseCatalogOperations {
   @Override
   public boolean dropTable(NameIdentifier ident) {
     try {
+      TableEntity tableEntity = store.get(ident, Entity.EntityType.TABLE, TableEntity.class);
+      Map<String, String> lancePropertiesMap = tableEntity.getProperties();
       String location = lancePropertiesMap.get("location");
-      // Remove the directory on storage
+
+      // Remove the table entity from the metastore
+      if (!store.delete(ident, Entity.EntityType.TABLE)) {
+        throw new RuntimeException("Failed to drop Lance table: " + ident.name());
+      }
+
+      // Remove the directory on storage, which deletes all data files. For cloud storage
+      // we need to the following code can work properly.
       FileSystem fs = FileSystem.get(new Configuration());
       return fs.delete(new Path(location), true);
     } catch (IOException e) {
